requirements.txt
fastapi
pydantic
sqlalchemy
python-magic
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
uvicorn
alembic 
pymysql 

